rss_reader_bot/requirements.txt
feedparser
python-telegram-bot==13.7
